feat(env): Added setup/mgmt scripts for the env
This commit brings a workshop.sh script for managing a testing harness
within the test lab VM. This will be leveraged throughout the workshop
to facilitate testing the various XDP programs.

There is also a workshop.env and common.sh added here to help with
managing the workshop.sh scripts functionality.