feat(build): Added in helper libs and makerules
This adds a makerules file that will be leveraged as an abstraction for
the various XDP program and user space programs to be built from local
Makefiles.

This also adds a set of helper headers directly pulled from the linux
kernel sources for v5.0. In absolutely no way is this my code or code
from NS1 and all rights and credit goes to the linux kernel developers.