feat(build): Cleaned up the Makefiles a bit
This cleans up the overall build process for the repo, specifically
adding in proper handling of the overall and local common headers for
each section when changes occur. As well as improves how the libbpf
build is triggered and managed, by moving its output to a dedicated
directory that can be ignored.